# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CommonResponse:
    properties:
      data:
        $ref: '#/definitions/dto.Resp'
    type: object
...
  dto.FeedDataResponseSwag:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.FeedSwag'
        type: array
    type: object
...
  dto.FeedResponseSwag:
    properties:
      data:
        $ref: '#/definitions/dto.FeedSwag'
    type: object
  dto.FeedSwag:
    properties:
      content:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      date:
        type: string
      id:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ImagesResponse:
    properties:
      keys:
        items:
          type: string
        type: array
    type: object
...
  dto.Resp:
    properties:
      success:
        type: boolean
    type: object
...
  models.IncrementRequest:
    properties:
      postIds:
        items:
          type: integer
        type: array
    type: object
...
  models.PostUpdateRequest:
    type: object
...
  s3.DeleteFilesRequest:
    properties:
      keys:
        items:
          type: string
        type: array
    type: object
  s3.DeleteFilesResponse:
    properties:
      error:
        type: string
      notFound:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
host: dev.lootor.me
info:
  contact: {}
  description: Документация к API Lootor.me
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lootor
  version: "1.0"
paths:
...
  /public/images/{key}:
    get:
      parameters:
      - description: id
        in: path
        name: key
        required: true
        type: string
      produces:
      - image/png
      - image/jpeg
      - image/webp
      responses:
        "200":
          description: image
          schema:
            type: file
      summary: получить изображение
      tags:
      - images
...
  /public/posts:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      - description: login
        in: query
        name: login
        required: true
        type: string
      - description: is draft
        in: query
        name: isDraft
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedDataResponseSwag'
      summary: получить посты юзера
      tags:
      - posts
  /public/posts/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedResponseSwag'
      summary: получить пост по id
      tags:
      - posts
  /public/posts/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        required: true
        type: string
      - description: order
        in: query
        name: order
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedDataResponseSwag'
      summary: все посты с пагинацией
      tags:
      - posts
  /public/posts/translit/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: translit
        in: path
        name: translit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedResponseSwag'
      summary: получить пост по id
      tags:
      - posts
...
  /public/thumbs/{key}:
    get:
      parameters:
      - description: имя файла
        in: path
        name: filename
        required: true
        type: string
      - description: ширина
        in: query
        name: width
        type: integer
      - description: высота
        in: query
        name: height
        type: integer
      - description: качество (1-100)
        in: query
        name: quality
        type: integer
      produces:
      - image/webp
      responses:
        "200":
          description: image
          schema:
            type: file
      summary: получить миниатюру изображения
      tags:
      - images
...
  /secured/images/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: поля
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/s3.DeleteFilesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/s3.DeleteFilesResponse'
      summary: удалить изображения из хранилища
      tags:
      - images
  /secured/images/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - collectionFormat: csv
        description: файлы для загрузки
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      - description: ширина картинки
        in: query
        name: width
        type: string
      - description: высота
        in: query
        name: height
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImagesResponse'
      summary: Загрузить изображения в хранилище
      tags:
      - images
...
  /secured/posts/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommonResponse'
      summary: удаление posta
      tags:
      - posts
    put:
      consumes:
      - application/json
      parameters:
      - description: поля
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/models.PostUpdateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedResponseSwag'
      summary: обновить контент поста
      tags:
      - posts
  /secured/posts/react/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: reaction
        in: query
        name: reaction
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedResponseSwag'
      summary: поставить реакцию
      tags:
      - posts
  /secured/posts/react/delete/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: reaction
        in: query
        name: reaction
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedResponseSwag'
      summary: убрать реакцию
      tags:
      - posts
  /secured/posts/views:
    post:
      consumes:
      - application/json
      parameters:
      - description: поля
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/models.IncrementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FeedResponseSwag'
      summary: просмотры
      tags:
      - posts
...
schemes:
- https
swagger: "2.0"